requirements.txt
python-dotenv
telethon
jinja2